% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dbWriteTable}
\alias{dbWriteTable}
\title{Write a \code{faketable} to a database table}
\usage{
dbWriteTable(src, name, reactive_faketable)
}
\arguments{
\item{src}{A DBIConnection object produced by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}}

\item{name}{A table name for a table present in the \code{src}}

\item{reactive_faketable}{A \link[shiny:reactive]{shiny::reactive} object that holds an underlying
\code{\link[=faketable]{faketable()}}}
}
\value{
\code{dbWriteTable} does not return, but does reassign the \code{faketable}
reactive object in the parent environment
}
\description{
Write a \code{faketable} to a database table
}
