% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.better_rbind}
\alias{.better_rbind}
\title{Bind multiple data frames by row}
\usage{
.better_rbind(..., .id = NULL)
}
\arguments{
\item{...}{Data frames to combine. Each argument can either be a data frame,
a list that could be a data frame, or a list of data frames. Columns are
matched by name, and any missing columns will be filled with \code{NA}.}

\item{.id}{The name of an optional identifier column. Provide a string to
create an output column that identifies each input. The column will use
names if available, otherwise it will use positions.}
}
\value{
A data frame the same type as the first element of \code{...}.
}
\description{
Bind any number of data frames by row, making a longer result. This is
similar to \code{do.call(rbind, dfs)}, but the output will contain all columns
that appear in any of the inputs.
}
\details{
Unlike \code{\link[dplyr:bind_rows]{dplyr::bind_rows()}}, \code{.better_rbind()} does not complain if
only one data frame is supplied.
}
\examples{
\dontrun{
.better_rbind(mtcars)
.better_rbind(mtcars, mtcars, .id = 'id')
}

}
\keyword{internal}
