% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.create_rowid}
\alias{.create_rowid}
\title{Add a \code{.rowId} column to a data frame}
\usage{
.create_rowid(data, rowId = NULL)
}
\arguments{
\item{data}{A data frame}

\item{rowId}{The name of the primary key column}
}
\value{
\code{data} with a new column named \code{.rowId} at column position one that
has an MD5 hash uniquely identifying that row
}
\description{
Add a \code{.rowId} column to a data frame
}
\examples{
faketables:::.create_rowid(mtcars)
mtcars |>
  tibble::rownames_to_column() |>
  faketables:::.create_rowid(rowId = 'rowname')

}
\keyword{internal}
