% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.reconstruct_inputs}
\alias{.reconstruct_inputs}
\title{Reconstruct a table from shiny inputs}
\usage{
.reconstruct_inputs(faketable, input)
}
\arguments{
\item{faketable}{A \code{\link[=faketable]{faketable()}} object}

\item{input}{The shiny server \code{input}}
}
\value{
The state of \code{faketable@data} as the user sees it in the UI
}
\description{
Reconstruct a table from shiny inputs
}
\details{
This is used internally to capture the changes made to the data in
the UI by the user. It first captures all of \code{input}, then filters by input
name to only inputs created by the table. These are then reconstructed into
a format matching \code{faketable@data} with column class handling done by the
user supplied \code{cast} function in \code{\link[=col_def]{col_def()}}.
}
\keyword{internal}
