% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-faketable.R
\name{faketable}
\alias{faketable}
\title{Create a \code{faketable} object}
\usage{
faketable(data, table_def, rowId = NULL, show_delete = FALSE)
}
\arguments{
\item{data}{A data.frame}

\item{table_def}{A \code{\link[=table_def]{table_def()}} object}

\item{rowId}{A character vector of length one identifying which column is a
primary key, if any.}

\item{show_delete}{A named list passed to the \code{\link[shiny:column]{shiny::column()}} that holds
the Delete button. If \code{FALSE}, the delete column will not be shown.}
}
\value{
A \code{faketable} object with the following properties:
\itemize{
\item \code{data}: The current state of the table of inputs
\item \code{inserted}: Rows from \code{data} that were inserted into the data (were not
present in \code{.raw_data})
\item \code{updated}: Rows from \code{data} that have been modified, but were present in
\code{.raw_data}
\item \code{deleted}: Rows that were in \code{.raw_data}, but have been removed from and
do not appear in \code{data}
\item \code{.raw_data}: The data originally pased to \code{\link[=faketable]{faketable()}} with
the addition of a \code{.rowId} column as the first column. This column is
calculated using by hashing either the provided \code{rowId} column or using the
row number and system time.
\item \code{.rowId}: The value of the \code{rowId} argument
\item \code{.data}: The same as \code{data}, but retains the \code{.rowId} column
\item \code{.inserted}: The same as \code{inserted}, but retains the \code{.rowId} column
\item \code{.deleted}: All rows that have been removed from \code{data}, including those that
were inserted then deleted
\item \code{.table_def}: A copy of the user supplied \code{\link[=table_def]{table_def()}} passed
as the argument \code{table_def}
\item \code{.show_delete}: A copy of the user supplied list passed as the argument
\code{show_delete}. If \code{TRUE} was supplied, it is replaced with \code{list()}.
\item \code{.iteration}: INTERNAL USE ONLY.
}
}
\description{
Create a \code{faketable} object
}
\details{
A \code{faketable} object is an \code{\link[S7:S7_object]{S7::S7_object()}} with the class
\code{faketable}. S7 object properties are accessed using an \code{@}, rather than
the traditional \code{$}. For example, the property \code{data} for a \code{faketable}
object called \code{faketable} can be accessed using \code{faketable@data}.

A note on \code{faketables@.iteration}: Shiny does not currently support
removing inputs from the server. This allows \code{faketables} to refresh fully
after using \code{\link[=dbWriteTable]{dbWriteTable()}}, by creating a new backend set of
input ids which otherwise would be duplicated because the data is the same.
}
\seealso{
For more details on \code{S7}, see the vignette \href{https://rconsortium.github.io/S7/articles/S7.html}{on the website} or by running:
\code{vignette('S7', package = 'S7')}
}
