% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faketables-package.R
\docType{package}
\name{faketables-package}
\alias{faketables}
\alias{faketables-package}
\title{faketables: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
