% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{input_call}
\alias{input_call}
\title{Create a \code{faketable} bare input call}
\usage{
input_call(fun, args)
}
\arguments{
\item{fun}{A bare \link{shiny} input function}

\item{args}{A list of named arguments to pass to \code{fun}. This should not
include an \code{inputId}.}
}
\value{
A \code{\link[=input_call]{input_call()}} object
}
\description{
Create a \code{faketable} bare input call
}
\examples{
input_call(shiny::textInput, args = list(label = 'Text Input'))
}
