% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{.render_table}
\alias{.render_table}
\alias{.create_table_body}
\title{Create and render a \code{faketables} table body}
\usage{
.render_table(f_tab, ns)

.create_table_body(f_tab, ns)
}
\arguments{
\item{f_tab}{A \code{\link[=faketable]{faketable()}} object}

\item{ns}{The session namespace from \code{shiny::NS()}or \code{session$ns}}
}
\value{
A \code{\link[shiny:fluidPage]{shiny::fluidRow()}} containing a \code{\link[shiny:column]{shiny::column()}} for each column
specified in \code{f_tab}'s \code{\link[=table_def]{table_def()}}
}
\description{
This output should be passed to \code{\link[shiny:renderUI]{shiny::renderUI()}} in the
server
}
\keyword{internal}
