% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-methods.R
\name{shiny-manipulate}
\alias{shiny-manipulate}
\alias{faketablesInsert}
\alias{faketablesUpdate}
\alias{faketablesDelete}
\title{Manipulate a \code{faketable} in Shiny}
\usage{
faketablesInsert(inputId = "faketables", reactive_faketable, data)

faketablesUpdate(inputId = "faketables", reactive_faketable, data)

faketablesDelete(inputId = "faketables", reactive_faketable, data)
}
\arguments{
\item{inputId}{An ID string used to identify the module UI}

\item{reactive_faketable}{A \link[shiny:reactive]{shiny::reactive} object that holds an underlying
\code{\link[=faketable]{faketable()}}}

\item{data}{A data.frame to add to the data in the \code{\link[=faketable]{faketable()}}
object. If it does not already have a primary key column as specified in
\code{\link[=table_def]{table_def()}}, one will be created and primary keys will be
generated.}
}
\value{
These methods do not return, but do reassign the \code{faketable}
reactive object in the parent environment
}
\description{
These are convenient references to \link{methods} for use
in Shiny applications.
}
\seealso{
For more details, see the vignette by running
\code{vignette('inserting_data')}
}
\keyword{internal}
