% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{table_def}
\alias{table_def}
\title{Create a \code{faketable} table definition}
\usage{
table_def(...)
}
\arguments{
\item{...}{At least one \code{\link[=col_def]{col_def()}} object. Only columns with a
definition will be displayed. If a column does not have a definition and
rows are added, any column without a definition will receive an \code{NA} value.}
}
\value{
A \code{\link[=table_def]{table_def()}} object
}
\description{
Create a \code{faketable} table definition
}
\examples{
\dontshow{if (rlang::is_installed('shinyjs')) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# to create a table_def to display mtcars$mpg
# see faketables::col_def() for more information
table_def(
 col_def(
   name = 'mpg',
   input = input_call(
     fun = shiny::textInput,
     args = list(label = NULL, placeholder = 'mpg')
   ),
   cast = as.numeric,
   width = 3,
   display_name = 'MPG'
 )
)
\dontshow{\}) # examplesIf}
}
