% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation}
\alias{validation}
\alias{is_input_call}
\alias{is_col_def}
\alias{is_table_def}
\alias{is_faketable}
\title{Check if an object is the correct class}
\usage{
is_input_call(x)

is_col_def(x)

is_table_def(x)

is_faketable(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical indication whether \code{x} belongs to the given class
}
\description{
Check if an object is the correct class
}
\keyword{internal}
