% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-faketable.R
\name{faketable}
\alias{faketable}
\title{Create a \code{faketable} object}
\usage{
faketable(x, table_def, rowId = "rowId", show_delete = list())
}
\arguments{
\item{x}{A data.frame}

\item{table_def}{A \code{\link[=table_def]{table_def()}} object}

\item{rowId}{A character vector of length one identifying which column to use
as a primary key}

\item{show_delete}{A named list passed to the \code{\link[shiny:column]{shiny::column()}} that holds
the Delete button. If \code{NULL}, the delete column will not be shown.}
}
\value{
A \code{faketable} object
}
\description{
Create a \code{faketable} object
}
