% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faketables-package.R
\docType{package}
\name{faketables-package}
\alias{faketables}
\alias{faketables-package}
\title{faketables: Create a Table of Shiny Inputs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create an interactive table of Shiny inputs by providing data and a table definition.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/landeranalytics/faketables}
  \item \url{https://landeranalytics.github.io/faketables/}
  \item Report bugs at \url{https://github.com/landeranalytics/faketables/issues}
}

}
\author{
\strong{Maintainer}: Gus Lipkin \email{gus@landeranalytics.com}

Other contributors:
\itemize{
  \item Lander Analytics [copyright holder, funder]
}

}
\keyword{internal}
