% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{methods}
\alias{methods}
\alias{insert}
\alias{update}
\alias{delete}
\title{\code{insert}, \code{update}, or \code{delete} rows from a \code{faketables} object}
\usage{
insert(f_tab, x)

update(f_tab, x)

delete(f_tab, x)
}
\arguments{
\item{f_tab}{A \code{\link[=faketable]{faketable()}} object}

\item{x}{\itemize{
\item \code{insert}: A data.frame to add to the data in the \code{\link[=faketable]{faketable()}}
object. If it does not already have a primary key column as specified in
\code{\link[=table_def]{table_def()}}, one will be created and primary keys will be
generated.
\item \code{update}: A data.frame with a primary key column as specified in
\code{\link[=table_def]{table_def()}} with primary key values already present in the
data.
\item \code{delete}: Either a character vector of values from the primary key column
as specified in \code{\link[=table_def]{table_def()}} or a data.frame of rows to remove
with that vector as a column.
}}
}
\value{
A \code{\link[=faketable]{faketable()}} object
}
\description{
\code{insert}, \code{update}, or \code{delete} rows from a \code{faketables} object
}
\details{
Unlike \code{\link[=update]{update()}} and \code{\link[=delete]{delete()}} which are
generally only used internally \code{\link[=insert]{insert()}} must be implemented by
end users. To do so, the return from \code{\link[=insert]{insert()}} must be passed
back into a \code{\link[=faketablesServer]{faketablesServer()}} call and then reassigned to the
users chosen variable as done with the initial
\code{\link[=faketablesServer]{faketablesServer()}} call.
}
\examples{
\dontrun{
# insert
# to insert new rows of data from a data.frame called `ins`
f_tab <- faketablesServer(faketable = insert(f_tab(), ins))

# update
# to update with a new column called 'new' that has the value 'new'
faketable <- update(faketable, dplyr::mutate(faketable@x, 'new' = 'new'))

# delete
# to delete the first six rows of the data where the primary key column is `rowId`
rows_to_delete <- utils::head(faketable@x)
faketable <- delete(faketable, rows_to_delete$.rowId)
# OR
faketable <- delete(faketable, rows_to_delete)
}
}
\keyword{internal}
