% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{methods}
\alias{methods}
\alias{insert}
\alias{update}
\alias{delete}
\title{\code{insert}, \code{update}, or \code{delete} rows from a \code{faketables} object}
\usage{
insert(faketable, data)

update(faketable, data)

delete(faketable, data)
}
\arguments{
\item{faketable}{A \code{\link[=faketable]{faketable()}} object}

\item{data}{\itemize{
\item \code{insert}: A data.frame to add to the data in the \code{\link[=faketable]{faketable()}}
object. If it does not already have a primary key column as specified in
\code{\link[=table_def]{table_def()}}, one will be created and primary keys will be
generated.
\item \code{update}: A data.frame with a primary key column as specified in
\code{\link[=table_def]{table_def()}} with primary key values already present in the
data.
\item \code{delete}: Either a character vector of values from the primary key column
as specified in \code{\link[=table_def]{table_def()}} or a data.frame of rows to remove
with that vector as a column.
}}
}
\value{
A \code{\link[=faketable]{faketable()}} object
}
\description{
\code{insert}, \code{update}, or \code{delete} rows from a \code{faketables} object
}
\examples{
\dontrun{
# insert
# to insert new rows of data from a data.frame called `ins`
faketable <- faketablesServer(faketable = insert(faketable(), ins))

# update
# to update with a new column called 'new' that has the value 'new'
faketable <- update(faketable, dplyr::mutate(faketable@data, 'new' = 'new'))

# delete
# to delete the first six rows of the data where the primary key column is `rowId`
rows_to_delete <- utils::head(faketable@data)
faketable <- delete(faketable, rows_to_delete$.rowId)
# OR
faketable <- delete(faketable, rows_to_delete)
}
}
\keyword{internal}
