% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{.render_header}
\alias{.render_header}
\alias{.create_table_header}
\title{Create and render a \code{faketables} table header}
\usage{
.render_header(faketable)

.create_table_header(faketable)
}
\arguments{
\item{faketable}{A \code{\link[=faketable]{faketable()}} object}
}
\value{
\code{NULL}

A \code{\link[shiny:fluidPage]{shiny::fluidRow()}} containing the display names from \code{faketable}'s
\code{\link[=table_def]{table_def()}} each rendered in \code{\link[shiny:column]{shiny::column()}}
}
\description{
Create and render a \code{faketables} table header
}
\keyword{internal}
